Move scoreboard component API to root package
