+++ cardinal-components-scoreboard/src/main/java/dev/onyxstudios/cca/api/v3/scoreboard/TeamAddCallback.java
@@ -20,27 +20,25 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package dev.onyxstudios.cca.api.v3.scoreboard.component;
+package dev.onyxstudios.cca.api.v3.scoreboard;
 
-import org.jetbrains.annotations.ApiStatus;
+import net.fabricmc.fabric.api.event.Event;
+import net.fabricmc.fabric.api.event.EventFactory;
+import net.minecraft.scoreboard.Team;
 
 /**
- * Entrypoint getting invoked to register <em>static</em> team component factories.
- *
- * <p>The entrypoint is exposed as {@code cardinal-components-scoreboard} in the mod json and runs for any environment.
- * It usually executes right before the first {@link net.minecraft.scoreboard.Scoreboard} instance is created.
- *
- * @since 2.4.2
+ * The callback interface for receiving global team synchronization events.
  */
-@ApiStatus.Experimental
-public interface ScoreboardComponentInitializer {
+@FunctionalInterface
+public interface TeamAddCallback {
+    Event<TeamAddCallback> EVENT = EventFactory.createArrayBacked(TeamAddCallback.class, (t) -> {}, listeners -> (team) -> {
+        for (TeamAddCallback callback : listeners) {
+            callback.onTeamAdded(team);
+        }
+    });
+
     /**
-     * Called to register component factories for statically declared component types.
-     *
-     * <p><strong>The passed registry must not be held onto!</strong> Static component factories
-     * must not be registered outside of this method.
-     *
-     * @param registry a {@link ScoreboardComponentFactoryRegistry} for <em>statically declared</em> components
+     * Called when a team's data is sent to all players
      */
-    void registerScoreboardComponentFactories(ScoreboardComponentFactoryRegistry registry);
+    void onTeamAdded(Team team);
 }
