--- cardinal-components-scoreboard/src/main/java/dev/onyxstudios/cca/api/v3/scoreboard/event/ScoreboardSyncCallback.java
@@ -1,45 +1,0 @@
-/*
- * Cardinal-Components-API
- * Copyright (C) 2019-2020 OnyxStudios
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
- * OR OTHER DEALINGS IN THE SOFTWARE.
- */
-package dev.onyxstudios.cca.api.v3.scoreboard.event;
-
-import net.fabricmc.fabric.api.event.Event;
-import net.fabricmc.fabric.api.event.EventFactory;
-import net.minecraft.scoreboard.Scoreboard;
-import net.minecraft.server.network.ServerPlayerEntity;
-
-/**
- * The callback interface for receiving scoreboard synchronization events.
- */
-@FunctionalInterface
-public interface ScoreboardSyncCallback {
-    Event<ScoreboardSyncCallback> EVENT = EventFactory.createArrayBacked(ScoreboardSyncCallback.class, (p, t) -> {}, listeners -> (player, scoreboard) -> {
-        for (ScoreboardSyncCallback callback : listeners) {
-            callback.onScoreboardSync(player, scoreboard);
-        }
-    });
-
-    /**
-     * Called when a team's data is sent to a player
-     */
-    void onScoreboardSync(ServerPlayerEntity player, Scoreboard scoreboard);
-}
